Add generated tests for 57 (randomly chosen) user commands from section 1
These utilities are fetched from the list generated by "fetch_groff.sh".
It is to be noted that many of these tests are environment dependent and
will most likely fail on a different machine. These specific testcases
will be eliminated once the feature for running the tests is
implemented. For more details, refer the first step of the pseudocode
mentioned in this document:
https://lists.freebsd.org/pipermail/soc-status/2017-July/001094.html

The purpose for adding these tests is to identify new cases and possible
scope for improvement in test generation.